Apply second-iteration review follow-ups
A fresh review pass after the seven group commits surfaced four
follow-ups, all fixed here:

- PaimonGraphBuilder.buildTableSummary mislabeled Paimon kind=1 entries
  as Iceberg-style positional deletes. Paimon's `_KIND=1` is a manifest
  log entry recording file removal, not a separate delete-file type.
  posDeleteFileCount / eqDeleteFileCount stay 0 for Paimon; dataFileCount
  reflects ADD entries; DELETE entries remain visible via
  manifestEntryCount and deleteManifestCount.
- GraphLayoutService manifest comparator: null sentinels widened to
  Long.MAX_VALUE to match the post-H-7 Long field type. Tables with
  sequence numbers > 2^31 would otherwise sort before the null sentinel.
- NodeDetails.PropertiesEvolutionSection: `setA + setB` is a List with
  duplicates, double-reporting changes for keys present in both versions.
  Switched to `setA union setB`.
- GraphCanvas: dropped duplicate `kotlin.math.roundToInt` import.

REVIEW.md updated to reflect final state.

+++ src/main/kotlin/service/GraphLayoutService.kt
@@ -151,8 +151,8 @@ object GraphLayoutService {
             val ma = (a as? GraphNode.ManifestNode)?.data
             val mb = (b as? GraphNode.ManifestNode)?.data
             compareValuesBy(ma, mb,
-                { it?.sequenceNumber ?: Int.MAX_VALUE },
-                { it?.minSequenceNumber ?: Int.MAX_VALUE },
+                { it?.sequenceNumber ?: Long.MAX_VALUE },
+                { it?.minSequenceNumber ?: Long.MAX_VALUE },
                 { it?.addedSnapshotId ?: Long.MAX_VALUE },
                 { IcebergGraphBuilder.manifestContentRank(it?.content) },
                 { it?.manifestPath ?: "" }
